requirements.txt
streamlit>=1.28.0
yt-dlp>=2023.11.16
openai-whisper>=20231117
chromadb>=0.4.15
sentence-transformers>=2.2.2
langchain>=0.0.346
pytube>=15.0.0
ffmpeg-python>=0.2.0
torch>=2.0.0
requests>=2.31.0